More details in algebra.tex
Changes summary:

Added the new results lemma-characterize-localization-map, lemma-localization-product and lemma-double-localization-prime. The first one is used to prove the second and third ones, which in turn are used to give more hints at the reader in several other proofs. The most notable of these are lemma-total-ring-fractions-no-embedded-points and lemma-characterize-reduced-ring-normal. Personally I found their proofs I little bit difficult to parse (in the former, finding out the isomorphism Aᵢ≅R_𝔮ᵢ comes from; in the latter, with the sentence "each factor is of the form R/𝔭ᵢ with field of fractions R_𝔭ᵢ"). With the level of detail I added, I believe the proofs now really explain what is going on. Again, and as always, if you think there's no need to write none or some of these details, feel free to trim any/all.